% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_functions.R
\name{count_points}
\alias{count_points}
\title{Count points within each hexagon in a lattice}
\usage{
count_points(h, pts, outcol = "npoints")
}
\arguments{
\item{h}{Lattice of hexagons; an object of class \code{hexlattice}
as produced by \code{\link{make_hexagons}}.}

\item{pts}{Point locations. This can either be a spatial data frame
(\code{sf} object) or a two-column matrix or data frame of point
coordinates.}

\item{outcol}{The name of the column of point counts to add to the
\code{hexlattice} object's data frame. If there is an existing column in the data
frame with this name, the name of the new column will be adjusted using
\code{\link[base]{make.names}}.}
}
\value{
A copy of the input \code{hexlattice} object with point counts added
  to its data frame.
}
\description{
Given a lattice of hexagons and a set of points, this function counts how
many points lie within each hexagon. The lattice should be a spatial data
frame (\code{sf} object) while the point locations can be a spatial data
frame or a two-column matrix or plain data frame. If points are provided as a
spatial data frame, the function checks whether the coordinate reference
system matches that of the hexagons. If either layer does not have a
reference system defined this check is skipped, ie. point and hexagon
coordinates are assumed to be comparable.
}
