% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_functions.R
\name{create_geometries}
\alias{create_geometries}
\title{Add geometries to a hexagonal lattice}
\usage{
create_geometries(h, force = FALSE)
}
\arguments{
\item{h}{Lattice of hexagons; an object of class \code{hexlattice}
as produced by \code{\link{make_hexagons}}.}

\item{force}{If TRUE, create new geometries even when there are existing
ones present. Default is FALSE.}
}
\value{
A copy of the input lattice with geometries added.
}
\description{
This function checks if the lattice already has geometries and, if not,
creates them.
}
