% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_functions.R
\name{has_geometries}
\alias{has_geometries}
\title{Check if a hexlattice object has geometries}
\usage{
has_geometries(h)
}
\arguments{
\item{h}{Lattice of hexagons; an object of class \code{hexlattice}
as produced by \code{\link{make_hexagons}}.}
}
\value{
TRUE if geometries are present, FALSE otherwise.
}
\description{
Check if a hexlattice object has geometries
}
