% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_functions.R
\name{hex2xy}
\alias{hex2xy}
\title{Convert hexagonal coordinates to cartesian coordinates}
\usage{
hex2xy(qr, h)
}
\arguments{
\item{h}{The reference lattice: an object of class \code{hexlattice}.}

\item{xy}{A matrix or data frame of hexagon locations. Q (hex column) and
R (hex row) ordinates are assumed to be in the first two columns.}
}
\value{
A matrix of cartesian coordinates.
}
\description{
Convert hexagonal coordinates to cartesian coordinates
}
