% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_functions.R
\name{hex_area2side}
\alias{hex_area2side}
\title{Calculate hexagon side length given area}
\usage{
hex_area2side(A)
}
\arguments{
\item{A}{Area values for one or more hexagons.}
}
\value{
Side length values.
}
\description{
A hexagon with area A has side length \code{s = sqrt[ 2*A / (3*sqrt(3)) ]}
}
\seealso{
\code{\link{hex_side2area}} for the reverse operation.
}
