% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_functions.R
\name{hex_area2width}
\alias{hex_area2width}
\title{Calculate hexagon width for a given area}
\usage{
hex_area2width(A)
}
\arguments{
\item{A}{Area values for one or more hexagons.}
}
\value{
Width (X-dimension) values.
}
\description{
In this package, hexagons are always positioned to have two sides vertical.
Width is the distance between these two sides.
}
\seealso{
\code{\link{hex_width2area}} for the reverse operation.
}
