% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_functions.R
\name{hex_side2height}
\alias{hex_side2height}
\title{Calculate hexagon height given side length}
\usage{
hex_side2height(s)
}
\arguments{
\item{s}{Side length values for one or more hexagons.}
}
\value{
Height (Y-dimension) values.
}
\description{
In this package, hexagons are always positioned to have two sides vertical.
Height is thus the distance between the two vertices on the vertical
mid-line, and is equal to \code{2 * s} (side length).
}
\seealso{
\code{\link{hex_height2side}} for the reverse operation.
}
