% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_functions.R
\name{hex_side2width}
\alias{hex_side2width}
\title{Calculate hexagon width given side length}
\usage{
hex_side2width(s)
}
\arguments{
\item{s}{Side length values for one or more hexagons.}
}
\value{
Width (X-dimension) value(s).
}
\description{
In this package, hexagons are always positioned to have two sides vertical.
Width is the distance between these two sides, and is equal to
\code{sqrt(3) * s} (side length).
}
\seealso{
\code{\link{hex_width2side}} for the reverse operation.
}
