% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_functions.R
\name{hex_width2area}
\alias{hex_width2area}
\title{Calculate hexagon area for a given width}
\usage{
hex_width2area(w)
}
\arguments{
\item{w}{Width (X-dimension) values for one or more hexagons.}
}
\value{
Hexagon areas.
}
\description{
In this package, hexagons are always positioned to have two sides vertical.
Width is the distance between these two sides.
}
\seealso{
\code{\link{hex_area2width}} for the reverse operation.
}
