% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_functions.R
\name{hex_width2side}
\alias{hex_width2side}
\title{Calculate hexagon side length given width}
\usage{
hex_width2side(w)
}
\arguments{
\item{w}{Width values for one or more hexagons.}
}
\value{
Side length value(s).
}
\description{
In this package, hexagons are always positioned to have two sides vertical.
Width is the distance between these two sides, and is equal to
\code{sqrt(3) * s} (side length).
}
\seealso{
\code{\link{hex_side2width}} for the reverse operation.
}
