% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_functions.R
\name{make_centroids}
\alias{make_centroids}
\title{Create centroids for a hexagon lattice}
\usage{
make_centroids(w, xbnds, ybnds)
}
\arguments{
\item{w}{Hexagon width.}

\item{xbnds}{Vector of two values for the minimum (left) and maximum (right)
X ordinates of the bounding rectangle.}

\item{ybnds}{Vector of two values for the minimum (bottom) and maximum (top)
Y ordinates of the bounding rectangle.}
}
\value{
A two column matrix of centroid coordinates.
}
\description{
Adapted from code by Tim Keitt posted at:
https://stat.ethz.ch/pipermail/r-sig-geo/2007-March/001791.html
}
\examples{
# Create 1 square kilometre hexagons over a given area
width <- hex_area2width(1000000)
centroids <- make_hexagons(width, xbnds = c(700000, 750000), ybnds = c(6500000, 6600000))

}
