% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_functions.R
\name{make_hexagons}
\alias{make_hexagons}
\title{Create a lattice of hexagons}
\usage{
make_hexagons(w, xbnds, ybnds, quiet = FALSE)
}
\arguments{
\item{w}{Hexagon width.}

\item{xbnds}{Vector of two values for the minimum (left) and maximum (right)
X ordinates of the bounding rectangle.}

\item{ybnds}{Vector of two values for the minimum (bottom) and maximum (top)
Y ordinates of the bounding rectangle.}
}
\value{
A list (class \code{hexlattice}) with the following elements:
  \code{hwidth} - hexagon width (X dimension);
  \code{centroids} - a two-column matrix of hexagon centroids;
  \code{edges} - a spatial data frame of hexagon boundaries;
  \code{xbnds} - X bounds;
  \code{ybnds} - Y bounds;
}
\description{
Create a lattice of hexagons
}
