% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot.hexlattice}
\alias{plot.hexlattice}
\title{Quick plot of hexagon edges using ggplot}
\usage{
\method{plot}{hexlattice}(h)
}
\arguments{
\item{h}{Lattice of hexagons; an object of class \code{hexlattice}
as produced by \code{\link{make_hexagons}}.}
}
\value{
A ggplot object.
}
\description{
Quick plot of hexagon edges using ggplot
}
