% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_functions.R
\name{remove_geometries}
\alias{remove_geometries}
\title{Remove geometries from a hexagonal lattice}
\usage{
remove_geometries(h)
}
\arguments{
\item{h}{Lattice of hexagons; an object of class \code{hexlattice}
as produced by \code{\link{make_hexagons}}.}
}
\value{
A copy of the input lattice with geometries removed.
}
\description{
This function checks if the given \code{hexlattice} object has geometries
and, if so, removes them. They can be restored later using
\code{\link{create_geometries}}.
}
