% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_functions.R
\name{sample_raster}
\alias{sample_raster}
\title{Sample values from a raster at hexagon centroid locations}
\usage{
sample_raster(h, r, interp = TRUE, outcol = NULL)
}
\arguments{
\item{h}{Lattice of hexagons; an object of class \code{hexlattice}
as produced by \code{\link{make_hexagons}}.}

\item{r}{A raster object, either a single layer (\code{RasterLayer}) or
a multi-layer raster (\code{RasterStack}, \code{RasterBrick}).}

\item{interp}{If \code{TRUE} (default), raster values are sampled using
bilinear interpolation; if \code{FALSE} no interpolation is performed.}

\item{outcol}{The name of the column(s) of raster values to add to the
\code{hexlattice} object's data frame. For multi-layer rasters, this is the
root name, and columns will be named \code{outcol1, outcol2, ...}. The
default (\code{NULL}) means to use raster layer names as column names if
these are defined, otherwise use the name 'layer'. If any names clash with
existing columns in the data frame, the name of the new columns will be
adjusted using \code{\link[base]{make.names}}.}
}
\value{
A copy of the input \code{hexlattice} object with raster values added
  to its data frame.
}
\description{
Given a lattice of hexagons and a raster layer, this function samples the
raster value at each hexagon centroid. A check is done that the hexagon
lattice and the raster have the same coordinate reference system. If either
does not have a reference system defined this check is skipped, i.e. point
and hexagon coordinates are assumed to be comparable.
}
