% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{shapes}
\alias{shapes}
\title{Return the lattice data frame}
\usage{
shapes(h)
}
\arguments{
\item{h}{Lattice of hexagons; an object of class \code{hexlattice}
as produced by \code{\link{make_hexagons}}.}
}
\value{
A data frame. If the lattice has geometries this will be an
  \code{sf} data frame.
}
\description{
This is the preferred way of accessing the data frame of a \code{hexlattice}
object.
}
