% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_functions.R
\name{xy2hex}
\alias{xy2hex}
\title{Convert cartesian coordinates to hexagonal coordinates}
\usage{
xy2hex(xy, h)
}
\arguments{
\item{xy}{A matrix or data frame of point locations. X and Y ordinates are
assumed to be in the first two columns.}

\item{h}{The reference lattice: an object of class \code{hexlattice}.}
}
\value{
A matrix of hexagon (axial) coordinates.
}
\description{
Convert cartesian coordinates to hexagonal coordinates
}
