% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_functions.R
\name{make_hexagons}
\alias{make_hexagons}
\title{Create a lattice of hexagons}
\usage{
make_hexagons(w, xbnds, ybnds, lazy = TRUE, quiet = FALSE)
}
\arguments{
\item{w}{Hexagon width.}

\item{xbnds}{Vector of two values for the minimum (left) and maximum (right)
X ordinates of the bounding rectangle.}

\item{ybnds}{Vector of two values for the minimum (bottom) and maximum (top)
Y ordinates of the bounding rectangle.}

\item{lazy}{If TRUE, hexagon geometries (\code{sf} polygons) are not
created immediately to save time and space. The lattice data frame
(element 'shapes' in the returned \code{hexlattice} object) will just
contain hexagon IDs, centroid (X,Y) coordinates and axial (Q-R)
coordinates. If FALSE, geometries will be created for the returned
lattice.}

\item{quiet}{If TRUE, suppress console ouput about the details of the
lattice being created.}
}
\value{
A list (class \code{hexlattice}) with the following elements:
  \describe{
    \item{width}{Hexagon width (X dimension).}
    \item{side}{Hexagon side length.}
    \item{area}{Hexagon area.}
    \item{xbnds}{X bounds specified when creating the lattice.}
    \item{ybnds}{Y bounds specified when creating the lattice.}
    \item{shapes}{A spatial data frame (class \code{sf}) with hexagon
    edges as the geometry column, and columns for id (integer: 1-N
    starting at lower left and proceeding row-wise), centroid X (xc),
    centroid Y (yc), hexagon lattice column (q), hexagon lattice row
    (r).}
  }
}
\description{
Create a lattice of hexagons
}
